feat: lpunpack super.raw

+++ clean_super.sh
@@ -57,6 +57,14 @@ fi
 echo -e '\e[3;36m[+] Deleting super.img as unneeded\e[0m'
 rm "$ROOT_DIR/super.img"
 
+echo -e '\e[3;36m[+] Finally extracting super.raw\e[0m'
+mkdir "$ROOT_DIR/super"
+lpunpack "$ROOT_DIR/super.raw" "$ROOT_DIR/super" &> /dev/null
+if [[ $? != 0 ]]; then
+    echo -e '\e[1;31m[-] Failed to extract the archive\e[0m'
+    exit 1
+fi
+
 echo -e "\e[3;36m[+] Removing working directory\e[0m"
 rm -rf "$ROOT_DIR"
